src/@api/notification/inbox-setting/resolvers/notification-delete-inbox-settings.resolver.ts
import { NotificationInboxSetting } from '@api/graphql';
import { NotificationDeleteInboxSettingsHandler } from '@api/notification/inbox-setting';
import { Auth } from '@aurora/decorators';
import { Auditing, AuditingMeta, QueryStatement, Timezone } from '@aurorajs.dev/core';
import { Args, Mutation, Resolver } from '@nestjs/graphql';

@Resolver()
@Auth('notification.inboxSetting.delete')
export class NotificationDeleteInboxSettingsResolver
{
    constructor(
        private readonly handler: NotificationDeleteInboxSettingsHandler,
    ) {}

    @Mutation('notificationDeleteInboxSettings')
    async main(
        @Args('query') queryStatement?: QueryStatement,
        @Args('constraint') constraint?: QueryStatement,
        @Timezone() timezone?: string,
        @Auditing() auditing?: AuditingMeta,
    ): Promise<NotificationInboxSetting[]>
    {
        return await this.handler.main(
            queryStatement,
            constraint,
            timezone,
            auditing,
        );
    }
}
